requirements.txt
chainlit==0.5.2
langchain==0.0.230
openai==0.27.8
python-dotenv==1.0.0
Requests==2.31.0